requirements.txt
# Core scraping & parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data handling
python-dotenv>=1.0.0

# Vector store
chromadb>=0.5.0

# Embeddings + LLM (via Ollama HTTP API)
ollama>=0.2.0

# API backend
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# Evaluation & testing
pytest>=8.2.0

# Utilities
tqdm>=4.66.0
